add modern logo svg
The ioshaven logo has been modified for a modern design. It keeps
a familiar look while providing optimizations for flat designs.

Easter egg: The squares represent apps and the lines combined with
the bite mark is like a wave. This is because iOS Havens logo was
originally designed to be a buoy (for safety) hince the continued
sea theme.